Choose user with most recent update time as admin
The rationale here is that the user with the most up-to-date sync time
is *probably* the least likely one to have been removed from the org.

This might not catch all cases, but at least it is easy to reason about,
rather than relying on ActiveRecord to hand us users to pick from.