Move jwt key under the jwt-aware addons
In addition to

```yaml
addons:
  sauce_connect:
    ⋮
  jwt:
    secure: …
```

also allow

```yaml
addons:
  sauce_connect:
    jwt:
      secure: …
```

Then pass it downstream.

If jwt definition is given both ways, keep both.

The intention here is that, by keeping both, we allow a gradual
transition to the latter.
When `travis-build` sees the old-style configuration and display
a deprecation warning, and stop processing at some point in
the future, say, a month's time.