Don't count migrated jobs towards concurrency and don't enqueue them
I've already attempted it to fix the problem in 01139a4, but I only
fixed one code path, which is currently not used, so the changes didn't
go into effect. This commit applies the changes to the other code path
as well.

The reason we need to do it is because when we migrate jobs from .org to
.com they might still be in a state that is considered by gatekeeper. So
they might be started but they're running on .org not .com, thus we
shouldn't count them as running on .com. The same with new jobs - we
don't want to start them as they would start on .org anyway.

I decided to filter these jobs in code instead of in the database
because the number of jobs that we fetch from the database is relatively
small and if I had to do it in the DB I would have to always join jobs
with repositories, which actually might be slower.